+++ src/resources/views/records/edit.blade.php
@@ -1,9 +1,9 @@
 @extends('layouts.app')
 
-@section('header')
+@push('scripts')
 	<link href="https://cdnjs.cloudflare.com/ajax/libs/select2/4.0.3/css/select2.min.css" rel="stylesheet" />
 	<script src="https://cdnjs.cloudflare.com/ajax/libs/select2/4.0.3/js/select2.min.js"></script>
-@endsection
+@endpush
 
 @section('content')
 <div class="container">